+++ src/component/sub-modules/extension/src/extension/delegate-children.js
@@ -9,7 +9,7 @@ KISSY.add("component/extension/delegate-children", function (S, Event) {
         ie = S.Env.host.document.documentMode || UA.ie,
         Features = S.Features,
         Gesture = Event.Gesture,
-        isTouchSupported = Features.isTouchSupported();
+        isTouchEventSupported = Features.isTouchEventSupported();
 
     function DelegateChildren() {
     }
@@ -52,16 +52,20 @@ KISSY.add("component/extension/delegate-children", function (S, Event) {
 
         __bindUI: function () {
             var self = this,
-                events;
+                events = Gesture.start +
+                    " " + Gesture.end +
+                    " " + Gesture.tap;
 
-                events = Gesture.start + " " + Gesture.end + " " + Gesture.tap + " touchcancel ";
+            if (Gesture.cancel) {
+                events += ' ' + Gesture.cancel;
+            }
 
-                if (!isTouchSupported) {
-                    events += "mouseover mouseout contextmenu " +
-                        (ie && ie < 9 ? "dblclick " : "");
-                }
+            if (!isTouchEventSupported) {
+                events += "mouseover mouseout contextmenu " +
+                    (ie && ie < 9 ? "dblclick " : "");
+            }
 
-                self.get("el").on(events, handleChildMouseEvents, self);
+            self.get("el").on(events, handleChildMouseEvents, self);
         },
 
         /**