+++ src/event/sub-modules/dom/sub-modules/touch/src/touch.js
@@ -5,12 +5,28 @@
  */
 KISSY.add('event/dom/touch', function (S, EventDomBase, eventHandleMap, eventHandle) {
 
+    var Features = S.Features;
+
     function setupExtra(event) {
         setup.call(this, event);
         eventHandleMap[event].setup.apply(this, arguments);
     }
 
     function setup(event) {
+        var self = this,
+            style = self.style;
+        if (Features.isMsPointerSupported() && style) {
+            if (!self.__ks_touch_action) {
+                self.__ks_touch_action = style.msTouchAction;
+                self.__ks_user_select = style.msUserSelect;
+                style.msTouchAction = style.msUserSelect = 'none';
+            }
+            if (!self.__ks_touch_action_count) {
+                self.__ks_touch_action_count = 1;
+            } else {
+                self.__ks_touch_action_count++;
+            }
+        }
         eventHandle.addDocumentHandle(this, event);
     }
 
@@ -20,6 +36,19 @@ KISSY.add('event/dom/touch', function (S, EventDomBase, eventHandleMap, eventHan
     }
 
     function tearDown(event) {
+        var self = this,
+            style = self.style;
+        if (Features.isMsPointerSupported() && style) {
+            if (!self.__ks_touch_action_count) {
+                S.error('touch event error for ie');
+            }
+            self.__ks_touch_action_count--;
+            if (!self.__ks_touch_action_count) {
+                style.msUserSelect = self.__ks_user_select;
+                style.msTouchAction = self.__ks_touch_action;
+                self.__ks_touch_action = '';
+            }
+        }
         eventHandle.removeDocumentHandle(this, event);
     }