+++ src/event/sub-modules/dom/sub-modules/touch/src/touch/handle.js
@@ -7,14 +7,15 @@ KISSY.add('event/dom/touch/handle', function (S, DOM, eventHandleMap, Event, Ges
 
     var key = S.guid('touch-handle'),
         Features = S.Features,
-        MOVE_DELAY = 30,
-        touchEvents = {
-        };
+        isMsPointerSupported = Features.isMsPointerSupported(),
+        touchEvents = {};
 
     touchEvents[Gesture.start] = 'onTouchStart';
     touchEvents[Gesture.move] = 'onTouchMove';
     touchEvents[Gesture.end] = 'onTouchEnd';
-    touchEvents['touchcancel'] = 'onTouchEnd';
+    if (Gesture.cancel) {
+        touchEvents[Gesture.cancel] = 'onTouchEnd';
+    }
 
     function DocumentHandler(doc) {
 
@@ -27,11 +28,42 @@ KISSY.add('event/dom/touch/handle', function (S, DOM, eventHandleMap, Event, Ges
 
         self.init();
 
+        // normalize pointer event to touch event
+        self.touches = [];
+
     }
 
     DocumentHandler.prototype = {
 
         constructor: DocumentHandler,
+
+        addTouch: function (t) {
+            this.touches.push(t);
+        },
+
+        removeTouch: function (t) {
+            var touches = this.touches;
+            S.each(touches, function (tt, index) {
+                if (tt.pointerId ==
+                    t.pointerId) {
+                    touches.splice(index, 1);
+                    return false;
+                }
+                return undefined;
+            });
+        },
+
+        updateTouch: function (t) {
+            var touches = this.touches;
+            S.each(touches, function (tt, index) {
+                if (tt.pointerId ==
+                    t.pointerId) {
+                    touches[index] = t;
+                    return false;
+                }
+                return undefined;
+            });
+        },
 
         init: function () {
             var self = this,
@@ -48,22 +80,35 @@ KISSY.add('event/dom/touch/handle', function (S, DOM, eventHandleMap, Event, Ges
             var type = e.type,
                 notUp,
                 touchList;
-            if (Features.isTouchSupported()) {
+            if (Features.isTouchEventSupported()) {
+                touchList = (type == 'touchend' || type == 'touchcancel') ?
+                    e.changedTouches : e.touches;
+                if (touchList.length == 1) {
+                    e.which = 1;
+                    e.pageX = touchList.pageX;
+                    e.pageY = touchList.pageY;
+                }
                 return e;
+            }
+            else if (isMsPointerSupported) {
+                touchList = this.touches;
             } else {
                 if (type.indexOf('mouse') != -1 && e.which != 1) {
                     return undefined;
                 }
                 touchList = [e];
-                notUp = !type.match(/up$/i);
-                e.touches = notUp ? touchList : [];
-                e.targetTouches = notUp ? touchList : [];
-                e.changedTouches = touchList;
-                return e;
             }
+            notUp = !type.match(/(up|cancel)$/i);
+            e.touches = notUp ? touchList : [];
+            e.targetTouches = notUp ? touchList : [];
+            e.changedTouches = touchList;
+            return e;
         },
 
         onTouchMove: function (e) {
+            if (isMsPointerSupported) {
+                this.updateTouch(e.originalEvent);
+            }
             // no throttle! to allow preventDefault
             this.callEventHandle('onTouchMove', e);
         },
@@ -76,11 +121,17 @@ KISSY.add('event/dom/touch/handle', function (S, DOM, eventHandleMap, Event, Ges
                 h = eventHandle[e].handle;
                 h.isActive = 1;
             }
+            if (isMsPointerSupported) {
+                self.addTouch(event.originalEvent);
+            }
             self.callEventHandle('onTouchStart', event);
         },
 
         onTouchEnd: function (event) {
             this.callEventHandle('onTouchEnd', event);
+            if (isMsPointerSupported) {
+                this.removeTouch(event.originalEvent);
+            }
         },
 
         callEventHandle: function (method, event) {